Django/QuotesAPI/requirements.txt
dj-database-url==2.1.0
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0